move core model to bio-model
