+++ bio/core/build.gradle
@@ -43,10 +43,11 @@ dependencies {
     compile group: "org.intermine", name: "intermine-resources", version: System.getProperty("imVersion")
     compile group: "org.intermine", name: "intermine-objectstore", version: System.getProperty("imVersion") // idresolver needs db
     compile group: "org.intermine", name: "intermine-api", version: System.getProperty("imVersion")
+    compile group: "org.intermine", name: "bio-source-flymine-static", version: System.getProperty("flymineVersion"), transitive: false
+    compile group: "org.intermine", name: "bio-model", version: System.getProperty("bioVersion"), transitive: false
     provided fileTree(dir: 'libs', include: '*.jar') //biojava-1.7.1.jar not found in public repo, obo and bbop not available in the public maven repo
     testCompile group: "org.intermine", name: "intermine-integrate", version: System.getProperty("imVersion"), classifier: "testClasses"
-    testCompile group: "org.intermine", name: "bio-source-flymine-static", version: System.getProperty("imVersion")
-    bioModel group: "org.intermine", name: "bio-model", version: System.getProperty("imVersion"), transitive: false
+    bioModel group: "org.intermine", name: "bio-source-flymine-static", version: System.getProperty("flymineVersion"), transitive: false
     runtime fileTree(dir: 'libs', include: '*.jar') //antlr-2.7.6-caching.jar
 }
