+++ bio/postprocess/build.gradle
@@ -43,7 +43,6 @@ subprojects {
     }
 
     configurations {
-        bioModel
         mergeModel
     }
