+++ bio/postprocess/create-chromosome-locations/build.gradle
@@ -18,8 +18,9 @@ sourceSets {
 }
 
 dependencies {
-    compile group: 'org.intermine', name: 'bio-model', version: bioVersion //to read genomic_keyDefs.properties
+    bioModel group: "org.intermine", name: "bio-source-flymine-static", version: System.getProperty("flymineVersion"), transitive: false
     compile group: 'org.intermine', name: 'intermine-integrate', version: imVersion
+    compile group: "org.intermine", name: "bio-source-flymine-static", version: System.getProperty("flymineVersion"), transitive: false
 }
 
 processResources {
