+++ bio/postprocess/create-gene-flanking-features/build.gradle
@@ -17,9 +17,12 @@ sourceSets {
     }
 }
 
+configurations {
+    testModel
+}
+
 dependencies {
-    compile group: 'org.intermine', name: 'bio-model', version: bioVersion //to read genomic_keyDefs.properties
-    compile group: 'org.intermine', name: 'intermine-integrate', version: imVersion
+    compile group: 'org.intermine', name: 'bio-source-flymine-static', version: System.getProperty("flymineVersion"), transitive: false
 }
 
 processResources {
