+++ bio/postprocess/create-intergenic-region-features/build.gradle
@@ -17,13 +17,10 @@ sourceSets {
     }
 }
 
-configurations {
-    testModel
-}
-
 dependencies {
-    compile group: 'org.intermine', name: 'bio-model', version: bioVersion //to read genomic_keyDefs.properties
     compile group: 'org.intermine', name: 'intermine-integrate', version: imVersion
+    compile group: 'org.intermine', name: 'bio-model', version: bioVersion, transitive: false
+    bioModel group: 'org.intermine', name: 'bio-model', version: bioVersion, transitive: false
 }
 
 processResources {
