+++ bio/postprocess/create-references/build.gradle
@@ -21,11 +21,6 @@ configurations {
     testModel
 }
 
-dependencies {
-    testModel group: 'org.intermine', name: 'bio-source-flymine-static', version: "2.0.0", transitive: false
-    compile group: 'org.intermine', name: 'intermine-integrate', version: imVersion
-}
-
 processResources {
     from('.') { include ("*.properties")}
 }
