+++ bio/sources/gradle.properties
@@ -1,2 +1,5 @@
 systemProp.imVersion=2.0.+
 systemProp.bioVersion=2.0.+
+# some unit tests need additional classes that aren't in the core model.
+# we use flymine's model. don't touch!
+systemProp.flymineVersion=2.0.+
