+++ bio/sources/vcf/src/main/resources/vcf_additions.xml
@@ -15,6 +15,8 @@
    <class name="SNV" extends="SequenceAlteration" is-interface="true"/>
    <class name="Substitution" extends="SequenceAlteration" is-interface="true"/>
 
+
+
    <!-- this is just a sequence feature, not an alteration -->
    <class name="TandemRepeat" extends="SequenceAlteration" is-interface="true"/>
 
@@ -35,5 +37,5 @@
    <class name="ConsequenceType" is-interface="true">
      <attribute name="type" type="java.lang.String"/>
    </class>
-
+    <class name="Transcript" extends="SequenceFeature" is-interface="true"/>
 </classes>
