+++ intermine/model/src/main/java/org/intermine/metadata/ReferenceDescriptor.java
@@ -101,8 +101,8 @@ public class ReferenceDescriptor extends FieldDescriptor
 
         } else {
             throw new MetaDataException("Unable to find ClassDescriptor for '"
-                    + referencedType + "' in model while processing: " + cld.getName() + "."
-                    + name);
+                    + referencedType + "' in model while processing: '" + cld.getName() + "'."
+                    + name + " " + cld.getName());
         }
 
         // find ReferenceDescriptor for the reverse reference
