+++ plugin/gradle.properties
@@ -1,4 +1,2 @@
 systemProp.imVersion=2.0.+
 systemProp.bioVersion=2.0.+
-systemProp.bioSourcesVersion=2.0.+
-
