+++ plugin/src/main/resources/META-INF/gradle-plugins/biomodel-dbmodel.properties
@@ -1,0 +1,1 @@
+implementation-class=org.intermine.plugin.dbmodel.BioModelDBModelPlugin
